+++ crates/full-node/sov-stf-runner/src/prover_service/stf_info.rs
@@ -1,3 +1,5 @@
+use serde::de::DeserializeOwned;
+use serde::{Deserialize, Serialize};
 use sov_rollup_interface::da::DaSpec;
 use sov_rollup_interface::zk::{StateTransitionPublicData, StateTransitionWitness};
 
@@ -8,17 +10,22 @@ pub(crate) struct BlockProof<Address, Da: DaSpec, Root> {
 }
 
 /// Holds all the necessary data for the creation of a block zk-proof.
+#[derive(Serialize, Deserialize)]
+#[serde(bound = "StateRoot: Serialize + DeserializeOwned, Witness: Serialize + DeserializeOwned")]
 pub struct StateTransitionInfo<StateRoot, Witness, Da: DaSpec> {
     /// Public input to the per block zk proof.
     pub(crate) data: StateTransitionWitness<StateRoot, Witness, Da>,
     /// Slot number.
-    pub(crate) slot_number: u64,
+    pub(crate) rollup_height: u64,
 }
 
 impl<StateRoot, Witness, Da: DaSpec> StateTransitionInfo<StateRoot, Witness, Da> {
     /// StateTransitionInfo constructor.
-    pub fn new(data: StateTransitionWitness<StateRoot, Witness, Da>, slot_number: u64) -> Self {
-        Self { data, slot_number }
+    pub fn new(data: StateTransitionWitness<StateRoot, Witness, Da>, rollup_height: u64) -> Self {
+        Self {
+            data,
+            rollup_height,
+        }
     }
 
     pub(crate) fn da_block_header(&self) -> &Da::BlockHeader {